requirements.txt
django-suit==0.2.15
DjangoUeditor